// web/bindings/Config.d.ts
// This file was automatically generated by ts_bind, do not modify it manually

import type { Action } from "./Action";
import type { Effect } from "./Effect";
import type { KnobAction } from "./KnobAction";


export interface Config {
   switchAction: [(Action)[], (Action)[], (Action)[], (Action)[], (Action)[], (Action)[]];
   knobAction: KnobAction;
   effect: Effect;
}